scripts/scraper/requirements.txt
scrapy==2.11.1
supabase==2.4.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
requests==2.31.0
slugify==0.0.1 